feat: :sparkles:

+++ src/controllers/vehicleController.js
@@ -1,19 +1,29 @@
 const db = require('../config/db');
+const request = require('request');
+require('dotenv').config();
 
-// Controlador para eliminar un vehículo por ID
 exports.deleteVehicle = async (req, res) => {
     const vehicleId = req.params.id;
 
     try {
-        // Eliminar el vehículo de la base de datos
-        const [result] = await db.execute('DELETE FROM Cars WHERE id = ?', [vehicleId]);
+        const userServiceUrl = `${process.env.QUERY_CAR_URL}/${vehicleId}`;
+        
+        request(userServiceUrl, { json: true }, async (err, response, body) => {
+            if (err) {
+                console.error(err);
+                return res.status(500).json({ message: 'Error fetching user data', error: err.message });
+            }
 
-        // Verificar si se eliminó el vehículo
-        if (result.affectedRows === 0) {
-            return res.status(404).json({ message: 'Vehicle not found' });
-        }
+            const userId = body.userId;
 
-        res.status(200).json({ message: 'Vehicle deleted successfully' });
+            const [result] = await db.execute('DELETE FROM Cars WHERE id = ? AND userId = ?', [vehicleId, userId]);
+
+            if (result.affectedRows === 0) {
+                return res.status(404).json({ message: 'Vehicle not found' });
+            }
+
+            res.status(200).json({ message: 'Vehicle deleted successfully' });
+        });
     } catch (error) {
         console.error(error);
         res.status(500).json({ message: 'Error deleting vehicle', error: error.message });
